+++ CHANGELOG.md
@@ -34,11 +34,15 @@
 ## Release Candidate 31 (TBD)
 
 #### Additions
+* Added Armored Jetpack
 
 #### Changes
 
 #### Fixes
 * Fixed #3390
+* Fixed research issues for vanilla items, e.g. Trident or Totem of Undying
+* Fixed #3368
+* Fixed #1315
 
 ## Release Candidate 30 (31 Dec 2021)
 https://thebusybiscuit.github.io/builds/TheBusyBiscuit/Slimefun4/stable/#30